+++ src/js/actions/explorer.js
@@ -6,42 +6,12 @@ export const REQUEST_STATION_DATA = 'REQUEST_STATION_DATA'
 export const RECEIVE_STATION_DATA = 'RECEIVE_STATION_DATA'
 export const REQUEST_ANIMATION_DATA = 'REQUEST_ANIMATION_DATA'
 export const RECEIVE_ANIMATION_DATA = 'RECEIVE_ANIMATION_DATA'
-export const SET_STEP = 'SET_STEP'
-
-export const setStep = (id) => ({
-  type: SET_STEP,
-  id
-})
 
 // Animation data:
 
 const shouldFetchData = (dataset) => (
   !dataset.isFetching || !dataset.items
 )
-
-const receiveAnimationData = (data) => ({
-  type: RECEIVE_ANIMATION_DATA,
-  data
-})
-
-const requestAnimationData = () => ({
-  type: REQUEST_ANIMATION_DATA
-})
-
-const fetchAnimationData = () => dispatch => {
-  dispatch(requestAnimationData())
-  return request
-    .get('data/mapanimation.json')
-    .then(({ body }) => {
-      dispatch(receiveAnimationData(body))
-    })
-}
-
-export const fetchAnimationDataIfNeeded = () => (dispatch, getState) => {
-  if (shouldFetchData(getState().animation)) {
-    return dispatch(fetchAnimationData())
-  }
-}
 
 export const hideStationDetails = () => ({
   type: HIDE_STATION_DETAILS
@@ -51,8 +21,6 @@ const showStationDetails = (data) => ({
   type: SHOW_STATION_DETAILS,
   data
 })
-
-// Explorer data:
 
 const findStation = (data, id) => {
   return data.find(({ID}) => ID.toString() === id.toString())
