+++ src/js/actions/navigation.js
@@ -1,0 +1,6 @@
+export const SET_STEP = 'SET_STEP'
+
+export const setStep = (id) => ({
+  type: SET_STEP,
+  id
+})
